♻️ Refactors species detail route and navigation
Updates the species detail route to be nested under a general 'info' route.

This change streamlines navigation and improves the URL structure for accessing species details.
It removes the direct routing to 'species/:species_id' and nests it under '/info/:species_id', enhancing maintainability and future extensibility of the routing system.